+++ vector/src/main/java/im/vector/riotx/features/roomdirectory/RoomDirectoryActivity.kt
@@ -65,8 +65,8 @@ class RoomDirectoryActivity : VectorBaseActivity() {
                 .subscribe { navigation ->
                     when (navigation) {
                         is Navigation.Back           -> onBackPressed()
-                        is Navigation.CreateRoom     -> addFragmentToBackstack(CreateRoomFragment(), R.id.simpleFragmentContainer)
-                        is Navigation.ChangeProtocol -> addFragmentToBackstack(RoomDirectoryPickerFragment(), R.id.simpleFragmentContainer)
+                        is Navigation.CreateRoom     -> addFragmentToBackstack(R.id.simpleFragmentContainer, CreateRoomFragment::class.java)
+                        is Navigation.ChangeProtocol -> addFragmentToBackstack(R.id.simpleFragmentContainer, RoomDirectoryPickerFragment::class.java)
                         is Navigation.Close          -> finish()
                     }
                 }
@@ -80,7 +80,7 @@ class RoomDirectoryActivity : VectorBaseActivity() {
 
     override fun initUiAndData() {
         if (isFirstCreation()) {
-            addFragment(PublicRoomsFragment(), R.id.simpleFragmentContainer)
+            addFragment(R.id.simpleFragmentContainer, PublicRoomsFragment::class.java)
         }
     }