+++ vector/src/main/java/im/vector/riotx/features/roomdirectory/createroom/CreateRoomActivity.kt
@@ -23,7 +23,6 @@ import androidx.appcompat.widget.Toolbar
 import androidx.lifecycle.ViewModelProviders
 import com.airbnb.mvrx.viewModel
 import im.vector.riotx.R
-import im.vector.riotx.core.di.ScreenComponent
 import im.vector.riotx.core.extensions.addFragment
 import im.vector.riotx.core.platform.ToolbarConfigurable
 import im.vector.riotx.core.platform.VectorBaseActivity
@@ -49,14 +48,9 @@ class CreateRoomActivity : VectorBaseActivity(), ToolbarConfigurable {
 
     override fun initUiAndData() {
         if (isFirstCreation()) {
-            addFragment(CreateRoomFragment(), R.id.simpleFragmentContainer)
-
+            addFragment(R.id.simpleFragmentContainer, CreateRoomFragment::class.java)
             createRoomViewModel.setName(intent?.getStringExtra(INITIAL_NAME) ?: "")
         }
-    }
-
-    override fun injectWith(injector: ScreenComponent) {
-        injector.inject(this)
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {